requirements-python-ingest.txt
lancedb
pymupdf
pytesseract
pillow
requests